Fix a valgrind complaint
The mtu_last variable of the struct UTPSocket  was not explicitly
initialized, so Valgrind complained when its value was printed in
the log. Actually, this variable was automatically initialized to
zero with the instantiation of the struct UTPSocket.

mtu_floor and mtu_ceiling are not explicitly initilized to zero
after the instantiation of the UTPSocket, but in this case
Valgrind does not complain, since they are assigned with a value,
before printing.